requirements.txt
phidata
python-dotenv
yfinance
packaging
duckduckgo-search # Agent will be able to do Web-Search
fastapi
uvicorn
groq
openai
python-multipart
sqlalchemy
pgvector
psycopg[binary]
pypdf
pytesseract